[guide] Remove ambiguity lib vs. executable
This commit transforms the guide example into a regular Rust executable
so that people following the Basic Workflow in the Burn book are not
surprised that the guide source does not correspond to their project.

In order to maintain the possibility to use the guide source as a library
like it is the case in the custom-renderer example, a lib.rs file is
added with a comment at the top for users.

In order to maintain the possibility to execute the guide as a regular
example the examples/guide.rs has been changed to simply call cargo. The
downside is that a new executable has to be compiled.
