+++ examples/guide/src/lib.rs
@@ -1,3 +1,9 @@
+//
+// Note: If you are following the Burn Book guide this file can be ignored.
+//
+// This lib.rs file is added only for convenience so that the code in this
+// guide can be reused.
+//
 pub mod data;
 pub mod inference;
 pub mod model;
